Extract D-inf flow accumulation into dedicated module and update hydrology API
- Move D-inf flow accumulation from flow_accumulation.py into
  flow_accumulation_dinf.py with its own test file
- Update basin, watershed, stream_order, and stream_order_mfd to use
  D8 flow direction consistently
- Add flow_accumulation_dinf to public API and docs
- Simplify flow_accumulation.py to D8-only logic
